splitting of coap tests
