+++ testing/data.json
@@ -79,25 +79,15 @@
         "info": "status of AC2889/10 in cli",
         "data": "[ConnectType]                 ConnectType: Localcontrol\n[DeviceId]                    DeviceId: 00000000000000000000000000000000\n[ProductId]                   ProductId: 00000000000000000000000000000001\n[Runtime]                     Runtime: 13.1 hours\n[StatusType]                  StatusType: localcontrol\n[WifiVersion]                 WifiVersion: AWS_Philips_AIR@54.2\n[aqil]                        Light brightness: 25\n[aqit]                        Air quality notification threshold: 4\n[cl]                          Child lock: False\n[ddp]                         Used index: IAI\n[fltsts0]                     Pre-filter and Wick: clean in 149 hours\n[fltsts1]                     HEPA filter: replace in 4300 hours\n[fltsts2]                     Active carbon filter: replace in 10423 hours\n[fltt1]                       HEPA filter type: NanoProtect Filter Series 3 (FY2422)\n[fltt2]                       Active carbon filter type: none\n[iaql]                        Allergen index: 1\n[mode]                        Mode: manual\n[modelid]                     ModelId: AC2889/10\n[name]                        Name: Luftreiniger\n[om]                          Fan speed: 1\n[ota]                         Over the air updates: ck\n[pm25]                        PM25: 1\n[pwr]                         Power: ON\n[swversion]                   Version: 1.0.7\n[type]                        Type: AC2889\n[uil]                         Buttons light: 2\n"
     },
-    "coap-firmware": {
-        "device": "AC2889/10",
-        "info": "firmware of AC2889/10",
-        "data": "{\"swversion\": \"1.0.7\"}"
-    },
     "coap-firmware-cli": {
         "device": "AC2889/10",
         "info": "firmware of AC2889/10 in cli",
         "data": "Version: 1.0.7\nOver the air updates: ck\n"
     },
-    "coap-fltsts": {
-        "device": "AC2889/10",
-        "info": "filterstatus of AC2889/10",
-        "data": "{\"fltsts0\": 213, \"fltsts1\": 4364, \"fltsts2\": 10487, \"fltt1\": \"A3\", \"fltt2\": \"none\"}"
-    },
     "coap-fltsts-cli": {
         "device": "AC2889/10",
         "info": "filterstatus of AC2889/10",
-        "data": "Pre-filter and Wick: clean in 0 hours\nHEPA filter: replace in 4526 hours\nActive carbon filter: replace in 10648 hours\n"
+        "data": "Pre-filter and Wick: clean in 149 hours\nHEPA filter: replace in 4300 hours\nActive carbon filter: replace in 10423 hours\n"
     },
     "coap-status-err193": {
         "device": "AC2889/10",
